requirements.txt
Authlib==1.4.1
blinker==1.9.0
boto3==1.36.12
botocore==1.36.12
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
Flask==3.1.0
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pycparser==2.22
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
sql==2022.4.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
